+++ tests/api-resources/v1/boxes/actions.test.ts
@@ -253,11 +253,8 @@ describe('resource actions', () => {
   });
 
   // Prism tests are disabled
-  test.skip('replayRecordingGet: only required params', async () => {
-    const responsePromise = client.v1.boxes.actions.replayRecordingGet(
-      'c9bdc193-b54b-4ddb-a035-5ac0c598d32d',
-      { seconds: 0 },
-    );
+  test.skip('rewindDisable', async () => {
+    const responsePromise = client.v1.boxes.actions.rewindDisable('c9bdc193-b54b-4ddb-a035-5ac0c598d32d');
     const rawResponse = await responsePromise.asResponse();
     expect(rawResponse).toBeInstanceOf(Response);
     const response = await responsePromise;
@@ -268,18 +265,8 @@ describe('resource actions', () => {
   });
 
   // Prism tests are disabled
-  test.skip('replayRecordingGet: required and optional params', async () => {
-    const response = await client.v1.boxes.actions.replayRecordingGet(
-      'c9bdc193-b54b-4ddb-a035-5ac0c598d32d',
-      { seconds: 0 },
-    );
-  });
-
-  // Prism tests are disabled
-  test.skip('replayRecordingStart', async () => {
-    const responsePromise = client.v1.boxes.actions.replayRecordingStart(
-      'c9bdc193-b54b-4ddb-a035-5ac0c598d32d',
-    );
+  test.skip('rewindEnable', async () => {
+    const responsePromise = client.v1.boxes.actions.rewindEnable('c9bdc193-b54b-4ddb-a035-5ac0c598d32d');
     const rawResponse = await responsePromise.asResponse();
     expect(rawResponse).toBeInstanceOf(Response);
     const response = await responsePromise;
@@ -290,10 +277,8 @@ describe('resource actions', () => {
   });
 
   // Prism tests are disabled
-  test.skip('replayRecordingStop', async () => {
-    const responsePromise = client.v1.boxes.actions.replayRecordingStop(
-      'c9bdc193-b54b-4ddb-a035-5ac0c598d32d',
-    );
+  test.skip('rewindExtract', async () => {
+    const responsePromise = client.v1.boxes.actions.rewindExtract('c9bdc193-b54b-4ddb-a035-5ac0c598d32d', {});
     const rawResponse = await responsePromise.asResponse();
     expect(rawResponse).toBeInstanceOf(Response);
     const response = await responsePromise;
@@ -384,8 +369,8 @@ describe('resource actions', () => {
   });
 
   // Prism tests are disabled
-  test.skip('setting', async () => {
-    const responsePromise = client.v1.boxes.actions.setting('c9bdc193-b54b-4ddb-a035-5ac0c598d32d');
+  test.skip('settings', async () => {
+    const responsePromise = client.v1.boxes.actions.settings('c9bdc193-b54b-4ddb-a035-5ac0c598d32d');
     const rawResponse = await responsePromise.asResponse();
     expect(rawResponse).toBeInstanceOf(Response);
     const response = await responsePromise;
@@ -396,8 +381,8 @@ describe('resource actions', () => {
   });
 
   // Prism tests are disabled
-  test.skip('settingReset', async () => {
-    const responsePromise = client.v1.boxes.actions.settingReset('c9bdc193-b54b-4ddb-a035-5ac0c598d32d');
+  test.skip('settingsReset', async () => {
+    const responsePromise = client.v1.boxes.actions.settingsReset('c9bdc193-b54b-4ddb-a035-5ac0c598d32d');
     const rawResponse = await responsePromise.asResponse();
     expect(rawResponse).toBeInstanceOf(Response);
     const response = await responsePromise;
@@ -408,8 +393,8 @@ describe('resource actions', () => {
   });
 
   // Prism tests are disabled
-  test.skip('settingUpdate: only required params', async () => {
-    const responsePromise = client.v1.boxes.actions.settingUpdate('c9bdc193-b54b-4ddb-a035-5ac0c598d32d', {
+  test.skip('settingsUpdate: only required params', async () => {
+    const responsePromise = client.v1.boxes.actions.settingsUpdate('c9bdc193-b54b-4ddb-a035-5ac0c598d32d', {
       scale: 1,
     });
     const rawResponse = await responsePromise.asResponse();
@@ -422,8 +407,8 @@ describe('resource actions', () => {
   });
 
   // Prism tests are disabled
-  test.skip('settingUpdate: required and optional params', async () => {
-    const response = await client.v1.boxes.actions.settingUpdate('c9bdc193-b54b-4ddb-a035-5ac0c598d32d', {
+  test.skip('settingsUpdate: required and optional params', async () => {
+    const response = await client.v1.boxes.actions.settingsUpdate('c9bdc193-b54b-4ddb-a035-5ac0c598d32d', {
       scale: 1,
     });
   });